Mutation probabilities and new mutation function
Added a function that computes computes the mutation probability of selected agent based off their fitness scores. This ensures the strongest selected agent will not be mutated.

Revamped the mutation function. If it's a selected agent it will mutate with a probability defined in the above function, if it is a product of crossover it'll mutate according to the fixed probability defined in GA initialisation.